+++ pkg/providers/iam/create_iam_group.go
@@ -2,7 +2,6 @@ package iam
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"path/filepath"
 	"strconv"
@@ -14,13 +13,6 @@ import (
 	"github.com/akamai/cli/pkg/terminal"
 	"github.com/fatih/color"
 	"github.com/urfave/cli/v2"
-)
-
-var (
-	// ErrFetchingUsersWithinGroup is returned when fetching users within group fails
-	ErrFetchingUsersWithinGroup = errors.New("unable to fetch users within group")
-	// ErrFetchingRolesWithinGroup is returned when fetching roles within group fails
-	ErrFetchingRolesWithinGroup = errors.New("unable to fetch roles within group")
 )
 
 // CmdCreateIAMGroup is an entrypoint to create-iam group command