+++ pkg/providers/iam/testdata/iam_all/import.sh
@@ -7,3 +7,5 @@ terraform import akamai_iam_group.group_id_101 101
 terraform import akamai_iam_group.group_id_102 102
 terraform import akamai_iam_group.group_id_112 112
 terraform import akamai_iam_group.group_id_123 123
+terraform import akamai_iam_ip_allowlist.allowlist ""
+terraform import akamai_iam_cidr_blocks.cidr_blocks ""