+++ pkg/providers/iam/testdata/iam_allowlist/multiple/import.sh
@@ -1,0 +1,3 @@
+terraform init
+terraform import akamai_iam_ip_allowlist.allowlist ""
+terraform import akamai_iam_cidr_blocks.cidr_blocks ""